==> 1_gathering_data.ipynb <==
## Gathering Data

import pandas as pd
import praw
from psaw import PushshiftAPI
import time

reddit = praw.Reddit(client_id = 'y45p1UGhNyNxnA',     # this is that personal use script key
                     client_secret = 'JaLr3ExpFi_1Dyi2BxqG2dblycQ', # this is that secret key
                     user_agent = 'project_3',    # Whatever name you gave your application
                     username = 'a4444999',      # your Reddit user name
                     password = 'Lodz1966')      # your Reddit password

api = PushshiftAPI(reddit)

#godddamn it when you get home, delete the creepy and pasta csvs, rerun this and resave bc its all fucked!!!!!!

### r/creepypasta

topics_dict = { "title":[],
            "score":[],
            "id":[],
            "url":[], 
            "comms_num": [],
            "created": [],
            "body":[]
              }

counter = 0 # initializing counting variable
sleep_counter = 0 # initializing sleep_counter variable
for item in api.search_submissions(subreddit='creepypasta', limit=5_000):
    topics_dict["title"].append(item.title)
    topics_dict["score"].append(item.score)
    topics_dict["id"].append(item.id)
    topics_dict["url"].append(item.url)
    topics_dict["comms_num"].append(item.num_comments)
    topics_dict["created"].append(item.created)
    topics_dict["body"].append(item.selftext)
    counter +=1
    sleep_counter += 1
    if sleep_counter > 99:
        print(counter)
        time.sleep(3)
        sleep_counter = 0

creepy_df = pd.DataFrame(topics_dict)
creepy_df['body'].shape

creepy_df.to_csv(path_or_buf='data/creepy_df.csv', index=False)
#index=False Unnamed doesn't show!! 

creepy_df.head()

#creepy_df['body'].value_counts().sort_values(ascending=False)

creepy_df['body'].replace(to_replace = ['[deleted]', '[removed]'], 
                          value = '',
                          inplace=True)

#combine body and head 

creepy_df['creepypasta'] = creepy_df['title'] + ': ' + creepy_df['body']

creepy_df['taget_col'] = 0

creepy_df.columns

creepy_df = creepy_df.drop(columns = ['title', 'score', 'id', 'url', 'comms_num', 'created', 'body'])

creepy_df.shape

creepy_df.head()

### r/nosleep

topics_dict = { "title":[],
            "score":[],
            "id":[],
            "url":[], 
            "comms_num": [],
            "created": [],
            "body":[]
              }
counter = 0 # initializing counting variable
sleep_counter = 0 # initializing sleep_counter variable
for item in api.search_submissions(subreddit='nosleep', limit=5_000):
    topics_dict["title"].append(item.title)
    topics_dict["score"].append(item.score)
    topics_dict["id"].append(item.id)
    topics_dict["url"].append(item.url)
    topics_dict["comms_num"].append(item.num_comments)
    topics_dict["created"].append(item.created)
    topics_dict["body"].append(item.selftext)
    counter +=1
    sleep_counter += 1
    if sleep_counter > 99:
        print(counter)
        time.sleep(3)
        sleep_counter = 0

nosleep_df = pd.DataFrame(topics_dict)
nosleep_df['body'].shape

nosleep_df.head()

nosleep_df.to_csv(path_or_buf='data/nosleep_df.csv', index=False)

nosleep_df['body'].replace(to_replace = ['[deleted]', '[removed]'],
                     value = '',
                     inplace=True)

nosleep_df['no_sleep'] = nosleep_df['title'] + ': ' + nosleep_df['body']

nosleep_df['taget_col'] = 1

nosleep_df.columns

nosleep_df = nosleep_df.drop(columns=['title', 'score', 'id', 'url', 'comms_num', 'created', 'body'])

nosleep_df.shape

### Combine DataFrames

nosleep_df.head()

creepy_df.head()

clean_df = pd.concat([creepy_df, nosleep_df])

clean_df.reset_index(inplace=True)

clean_df = clean_df.drop(columns='index')

clean_df;

clean_df.to_csv(path_or_buf='data/clean_data.csv', index=False)
