==> 2_cleaning.ipynb <==
import pandas as pd
import numpy as np
import re

# Importing raw reddit data & removing special characters

creepy = pd.read_csv('data/creepy_df.csv')
nosleep = pd.read_csv('data/nosleep_df.csv')

creepy.isnull().sum();

nosleep.isnull().sum();

creepy = creepy.fillna('')
nosleep = nosleep.fillna('')

creepy['body'].replace(to_replace = ['[deleted]', '[removed]'], 
                          value = '',
                          inplace=True)

creepy['text'] = creepy['title'] + ': ' + creepy['body']
creepy['target'] = 0
creepy = creepy.drop(columns = ['title', 'score', 'id', 'url', 'comms_num', 'created', 'body'])


nosleep['body'].replace(to_replace = ['[deleted]', '[removed]'], 
                          value = '',
                          inplace=True)

nosleep['text'] = nosleep['title'] + ': ' + nosleep['body']
nosleep['target'] = 1
nosleep = nosleep.drop(columns = ['title', 'score', 'id', 'url', 'comms_num', 'created', 'body'])

nosleep.head();

nosleep.text = [re.sub('[^A-Za-z0-9!?]+', ' ', text) for text in nosleep.text]

creepy.text = [re.sub('[^A-Za-z0-9!?]+', ' ', text) for text in creepy.text]

nosleep.shape

creepy.shape

## checking for special characters

is_ascii_ns = nosleep['text'].apply(lambda s: len(s) == len(s.encode()))
nosleep = nosleep.loc[is_ascii_ns,:]

nosleep.shape

is_ascii_cp = creepy['text'].apply(lambda s: len(s) == len(s.encode()))
creepy = creepy.loc[is_ascii_cp,:]

creepy.shape

ascii_clean = pd.concat([creepy, nosleep])

ascii_clean.to_csv(path_or_buf='data/ascii_clean.csv', index=False)

ascii_clean
