==> 3_clean_data_exploration.ipynb <==
import pandas as pd
import numpy as np

from sklearn.linear_model import LogisticRegression
from sklearn.tree import DecisionTreeClassifier
from sklearn.ensemble import BaggingClassifier, RandomForestClassifier, ExtraTreesClassifier, GradientBoostingClassifier, AdaBoostClassifier
from sklearn.svm import SVC

from sklearn.feature_extraction.text import CountVectorizer, TfidfVectorizer, ENGLISH_STOP_WORDS

from nltk.corpus import stopwords

from sklearn.preprocessing import StandardScaler
from sklearn.model_selection import train_test_split, GridSearchCV
from sklearn.pipeline import Pipeline

import matplotlib.pyplot as plt
%matplotlib inline

### In this notebook I gridsearched tfidf and Logistic regression for the the data cleaned in 3 different ways. I found that the corpus that wasn't stripped of special characters gave the best score (df3)

df = pd.read_csv('data/ascii_clean.csv') 

X = df['text']
y = df['target']

X_train, X_test, y_train, y_test = train_test_split(X, y, 
                                                    test_size = 0.2,
                                                    random_state=42)

english_stop = list(ENGLISH_STOP_WORDS)

pipe_tvec = Pipeline(
    [('tf', TfidfVectorizer()),
     ('lr', LogisticRegression())]
    ) 

pipe_params = {'tf__max_features': [None],
               'tf__strip_accents': [None, 'ascii'], 
               'tf__stop_words': [stopwords.words("english"), english_stop],
               'tf__ngram_range': [(1,1), (1,2)],
               'tf__min_df': [0.005, 1],
             
               'lr__solver': ['saga'],
               'lr__penalty': ['l2'],
               'lr__n_jobs': [-1]
               }

gs = GridSearchCV(pipe_tvec,
                  param_grid=pipe_params, 
                  cv=5,
                  verbose=1) 

gs.fit(X_train, y_train)
print(gs.best_score_)
print(gs.best_params_)

df2 = pd.read_csv('rg_clean.csv')

X = df2['text']
y = df2['target_col']

X_train, X_test, y_train, y_test = train_test_split(X, y, 
                                                    test_size = 0.2,
                                                    random_state=42)

pipe_tvec = Pipeline(
    [('tf', TfidfVectorizer()),
     ('lr', LogisticRegression())]
    ) 

pipe_params = {'tf__max_features': [None],
               'tf__strip_accents': [None, 'ascii'], 
               'tf__stop_words': [stopwords.words("english"), english_stop],
               'tf__ngram_range': [(1,1), (1,2)],
               'tf__min_df': [0.005, 1],
             
               'lr__solver': ['saga'],
               'lr__penalty': ['l2'],
               'lr__n_jobs': [-1]
               }

gs = GridSearchCV(pipe_tvec,
                  param_grid=pipe_params, 
                  cv=5,
                  verbose=1) 

gs.fit(X_train, y_train)
print(gs.best_score_)
print(gs.best_params_)

df3 = pd.read_csv('data/clean_data.csv')

X = df3['text']
y = df3['target_col']

X_train, X_test, y_train, y_test = train_test_split(X, y, 
                                                    test_size = 0.2,
                                                    random_state=42)

pipe_tvec = Pipeline(
    [('tf', TfidfVectorizer()),
     ('lr', LogisticRegression())]
    ) 

pipe_params = {'tf__max_features': [None],
               'tf__strip_accents': [None, 'ascii'], 
               'tf__stop_words': [stopwords.words("english"), english_stop],
               'tf__ngram_range': [(1,1), (1,2)],
               'tf__min_df': [0.005, 1],
             
               'lr__solver': ['saga'],
               'lr__penalty': ['l2'],
               'lr__n_jobs': [-1]
               }

gs = GridSearchCV(pipe_tvec,
                  param_grid=pipe_params, 
                  cv=5,
                  verbose=1) 

gs.fit(X_train, y_train)
print(gs.best_score_)
print(gs.best_params_)

## d3

df3 = pd.read_csv('data/clean_data.csv')

X = df3['text']
y = df3['target_col']

X_train, X_test, y_train, y_test = train_test_split(X, y, 
                                                    test_size = 0.2,
                                                    random_state=42)

tf = TfidfVectorizer(stop_words=stopwords.words("english"),
                     max_features=None,
                     ngram_range=(1,1),
                     strip_accents='ascii' 
                    )

tf.fit(X_train)
X_train_tf = tf.transform(X_train).todense()
X_test_tf = tf.transform(X_test).todense()

X_train_tf = tf.transform(X_train) #.todense()
X_test_tf = tf.transform(X_test) 

lr = LogisticRegression(solver='saga')

lr.fit(X_train_tf, y_train)
(lr.score(X_train_tf, y_train), lr.score(X_test_tf, y_test))

# df2

df2 = pd.read_csv('rg_clean.csv')

X = df2['text']
y = df2['target_col']

X_train, X_test, y_train, y_test = train_test_split(X, y, 
                                                    test_size = 0.2,
                                                    random_state=42)

tf = TfidfVectorizer(stop_words=stopwords.words("english"),
                     max_features=None,
                     ngram_range=(1,1),
                     strip_accents='ascii' 
                    )

tf.fit(X_train)
X_train_tf = tf.transform(X_train).todense()
X_test_tf = tf.transform(X_test).todense()

X_train_tf = tf.transform(X_train) #.todense()
X_test_tf = tf.transform(X_test) 

lr = LogisticRegression(solver='saga')

lr.fit(X_train_tf, y_train)
(lr.score(X_train_tf, y_train), lr.score(X_test_tf, y_test))

# df

df = pd.read_csv('data/ascii_clean.csv') 

X = df['text']
y = df['target']

X_train, X_test, y_train, y_test = train_test_split(X, y, 
                                                    test_size = 0.2,
                                                    random_state=42)

tf = TfidfVectorizer(stop_words=stopwords.words("english"),
                     max_features=None,
                     ngram_range=(1,1),
                     strip_accents='ascii' 
                    )

tf.fit(X_train)
X_train_tf = tf.transform(X_train).todense()
X_test_tf = tf.transform(X_test).todense()

X_train_tf = tf.transform(X_train) #.todense()
X_test_tf = tf.transform(X_test) 

lr = LogisticRegression(solver='saga')

lr.fit(X_train_tf, y_train)
(lr.score(X_train_tf, y_train), lr.score(X_test_tf, y_test))
